+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 *Mar 29, 2026*
 
+- Fix `ThumbnailOptions` double-applying EXIF orientation when `createThumbnailWithTransform` is enabled – https://github.com/kean/Nuke/issues/870
 - Fix Xcode 26.0 compatibility – https://github.com/kean/Nuke/issues/871
 
 ## Nuke 13.0